Working plot without delta_R
