+++ src/config.py
@@ -3,6 +3,7 @@ import os
 CWD = os.getcwd()
 PILEUP_PATH = f"{CWD}/data/1-initial/pileup.csv"
 TT_PATH = f"{CWD}/data/1-initial/ttbar.csv"
+TT_INTER_PATH = f"{CWD}/data/2-intermediate/ttbar_jets.csv"
 
 MAX_DATA_ROWS = None
